make renderFilteredItems top-level export, add docs
